% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrifuge_tools.R
\name{add_cols}
\alias{add_cols}
\title{Add extra columns to a centrifuge table}
\usage{
add_cols(cent_out, cent_report)
}
\arguments{
\item{cent_out}{Centrifuge table}

\item{cent_report}{Centrifuge report table}
}
\description{
Add extra columns to a centrifuge table
}
