% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Add missing columns to a table}
\usage{
add_missing_columns(table, column_names, fill = NA)
}
\arguments{
\item{table}{A data.frame or matrix}

\item{column_names}{Character vector; List of column names to add}

\item{fill}{Value to fill the new columns}
}
\value{
data.frame or matrix
}
\description{
Add missing columns to a table
}
