% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_compare_all}
\alias{boot_compare_all}
\title{Apply a two group statistical test to each group in a nested tibble}
\usage{
boot_compare_all(grouped_mutation_table, test, within = NULL, reps = 1e+05)
}
\description{
Apply a two group statistical test to each group in a nested tibble
}
