% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_quantiles}
\alias{boot_quantiles}
\title{Bootstrap a statistic across a table and produce bca ci and quantiles}
\usage{
boot_quantiles(tbl, statistic, reps = 10000)
}
\arguments{
\item{tbl}{A tibble}

\item{statistic}{function; A bootstrap statistic}

\item{reps}{integer; Number of bootstrap replicates}
}
\description{
Bootstrap a statistic across a table and produce bca ci and quantiles
}
