% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{bootstrap_centered_null}
\alias{bootstrap_centered_null}
\title{Bootstrap the null distribution for a statistic centered at a given value}
\usage{
bootstrap_centered_null(mutation_table, statistic, center, ..., reps = 10000)
}
\description{
This function generates a distribution for the null hypothesis that two
groups do not differ in some statistic. For example if we were interested in
whether two samples had a different mean; this function would produce a
bootstrapped sample for one group centered at the actual mean across both
groups. This ensures that the bootstrapped distribution corresponds to the
actual null hypothesis. Explained better here:
https://stats.stackexchange.com/questions/136661
}
