% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{bootstrap_test}
\alias{bootstrap_test}
\title{Bootstrap difference in statistic across two groups}
\usage{
bootstrap_test(tbl, statistic, reps = 10000)
}
\arguments{
\item{statistic}{function; A valid bootstrap statistic}

\item{reps}{integer; Number of bootstrap replicates}

\item{mutation_table}{tibble; A mutation table nested by a group}
}
\value{
A p-value
}
\description{
See:
https://stats.stackexchange.com/questions/136661
Also:
Efron's and Tibshirani's - intro to the bootstrap, page 223
}
