% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{calc_bca_quantiles}
\alias{calc_bca_quantiles}
\title{Calculate quantiles and bias corrected confidence intervals for a boot object}
\usage{
calc_bca_quantiles(boot_object)
}
\description{
Calculate quantiles and bias corrected confidence intervals for a boot object
}
