% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{center_distribution}
\alias{center_distribution}
\title{Center a distribution at a given mean value}
\usage{
center_distribution(d, target)
}
\arguments{
\item{d}{numeric; the distribution}

\item{target}{numeric; the new mean}
}
\value{
The centered distribution
}
\description{
Center a distribution at a given mean value
}
