% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{extract_numeric}
\alias{extract_numeric}
\title{Grab all the numbers from a list as a string}
\usage{
extract_numeric(x)
}
\arguments{
\item{x}{A string}
}
\value{
Numeric vector
}
\description{
Grab all the numbers from a list as a string
}
