% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobtools_util.R
\name{extract_taxa}
\alias{extract_taxa}
\title{Remove all entries from blobtab except those from taxa in taxlist}
\usage{
extract_taxa(blobtab, taxlist)
}
\arguments{
\item{blobtab}{A table outputted from 'blobtools view'}

\item{A}{list of taxa to include}
}
\value{
A table with all other taxa removed
}
\description{
Remove all entries from blobtab except those from taxa in taxlist
}
