% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{gen_square_matrix}
\alias{gen_square_matrix}
\title{Generate an empty square matrix}
\usage{
gen_square_matrix(dimnames, fill = NA)
}
\arguments{
\item{dimnames}{character; The column/row dimensions.}
}
\value{
A matrix with dimnames as the column and row names.
}
\description{
Generate an empty square matrix
}
