% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{list_to_string}
\alias{list_to_string}
\title{Convert a list to a space separated string}
\usage{
list_to_string(lst)
}
\arguments{
\item{lst}{A list}
}
\value{
A string
}
\description{
Convert a list to a space separated string
}
