% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{precomputed_barplot}
\alias{precomputed_barplot}
\title{Barplot with precomputed confidence intervals}
\usage{
precomputed_barplot(df, xlab, ylab, xval, yscale = "unscaled", fill = NULL,
  limits = NULL, legend = TRUE)
}
\description{
Barplot with precomputed confidence intervals
}
