% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{precomputed_boxplot}
\alias{precomputed_boxplot}
\title{Box and whiskers plot with precomputed confidence intervals}
\usage{
precomputed_boxplot(df, xlab, ylab, xval, yscale = "unscaled", fill = NULL,
  limits = NULL, legend = TRUE)
}
\description{
Box and whiskers plot with precomputed confidence intervals
}
