% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_util.R
\name{read.text}
\alias{read.text}
\title{Read a text file as a single string}
\usage{
read.text(filename)
}
\arguments{
\item{filename}{File path}
}
\value{
Character
}
\description{
Read a text file as a single string
}
