% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{replace_rows}
\alias{replace_rows}
\title{Replace rows in a data.frame by index}
\usage{
replace_rows(df, old_index, new_row)
}
\arguments{
\item{df}{data.frame}

\item{old_index}{Numeric; Index of the old row in df}

\item{new_row}{Vector; New row to add to df}
}
\value{
data.frame
}
\description{
Replace rows in a data.frame by index
}
