% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{select_groups}
\alias{select_groups}
\title{Filter a data.frame to exclude all rows which are not in 'groups'}
\usage{
select_groups(df, by, groups)
}
\arguments{
\item{df}{A data.frame}

\item{by}{Character; A column name from 'df'}

\item{groups}{Factor; Vector of levels}
}
\value{
A data.frame
}
\description{
Filter a data.frame to exclude all rows which are not in 'groups'
}
