% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{split_nested_tibble}
\alias{split_nested_tibble}
\title{Split a nested tibble by a column}
\usage{
split_nested_tibble(tbl, by)
}
\arguments{
\item{tbl}{A nested tibble}

\item{by}{A column name within the nested tibble to split by}
}
\value{
A list of nested tibbles
}
\description{
Split a nested tibble by a column
}
