% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{tibble_combn}
\alias{tibble_combn}
\title{Apply a function across all combinations of levels of a tibble group}
\usage{
tibble_combn(dat, set_size, func, func_type = summarize,
  grouped_input = FALSE, within = NULL, ...)
}
\arguments{
\item{dat}{tbl_df A grouped tibble}

\item{set_size}{Number of levels in group to compare in each iteration}

\item{func}{function Function to apply to each set}

\item{func_type}{Type of dplyr operation, e.g do, summarize}

\item{grouped_input}{bool If TRUE, groups in 'dat' will be conserved in calls
to func}

\item{within}{character A column name from 'dat'. If given, comparisons will
only be made between values which share the same
values in the within column.}
}
\value{
A list of results from 'func'
}
\description{
All extra parameters are passed to the input function
}
