% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{unlist_preserving_names}
\alias{unlist_preserving_names}
\title{Unlist without changing the names of the inner objects}
\usage{
unlist_preserving_names(x, recursive = FALSE)
}
\description{
Unlist without changing the names of the inner objects
}
