% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_util.R
\name{write.table.with.header}
\alias{write.table.with.header}
\title{Write a table to file with added header using write.table}
\usage{
write.table.with.header(x, file, header, ...)
}
\arguments{
\item{x}{The table}

\item{file}{File path to write to}

\item{header}{The header}
}
\description{
Write a table to file with added header using write.table
}
