% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_table_with_header}
\alias{write_table_with_header}
\title{Write a table to file with added header using write.table}
\usage{
write_table_with_header(x, file, header, ...)
}
\arguments{
\item{x}{The table}

\item{file}{File path to write to}

\item{header}{The header}
}
\description{
Write a table to file with added header using write.table
}
