% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{tibble_combn}
\alias{tibble_combn}
\title{Apply a function across all combinations of groups in a nested tibble}
\usage{
tibble_combn(tbl, set_size, func, within = NULL, ...)
}
\arguments{
\item{tbl}{A nested tibble}

\item{set_size}{Number of rows in dat to compare in each iteration}

\item{func}{function Function to apply to each set of rows}

\item{within}{character A column name from the inner nested tibbles in `tbl`.
If given, comparisons will only be made between rows
which share the same group in the within column.}
}
\value{
A list of results from 'func'
}
\description{
All extra parameters are passed to the input function
}
